refactor: change style file names, change folder structure
