+++ src/commerce-tools/models/types/auth-state.ts
