--- src/components/label/_label.scss
@@ -1,2 +1,0 @@
-@use '../../styles/scss/variables';
-@use '../../styles/scss/mixins';
