+++ src/components/modals/confirm-modal/styles.scss
