+++ src/pages/header.ts
@@ -1,6 +1,6 @@
+import { authService } from '../commerce-tools/auth-service.ts';
 import { CreateButton } from '../components/button/create-button.ts';
 import { route } from '../router/index.ts';
-import { authService } from '../services/commercetools/auth-service.ts';
 import { CreateElement } from '../shared/utils/create-element.ts';
 
 export class Header {
