+++ src/pages/registration/personal-info-elements.ts
@@ -3,7 +3,7 @@ import { Label } from '../../components/label/label.ts';
 import type { IParameters } from '../../shared/models/interfaces/index.ts';
 import { CreateElement } from '../../shared/utils/create-element.ts';
 
-import './registration.scss';
+import './styles.scss';
 
 export class PersonalInfoElements extends CreateElement {
   public inputName: CreateInput;
