+++ src/pages/registration/styles.scss
