+++ src/tests/auth-service.test.ts
@@ -1,7 +1,7 @@
 import type { CustomerSignInResult } from '@commercetools/platform-sdk';
 import { describe, expect, test, vi } from 'vitest';
 
-import { AuthorizationService, authService } from '../services/commercetools/auth-service';
+import { AuthorizationService, authService } from '../commerce-tools/auth-service';
 
 describe('Check for singleton', () => {
   test('should return instance if this instance doesn`t exist', () => {
